requirements.txt
requests>=2.31.0
pyngrok>=7.0.0
pyinstaller>=6.0.0
customtkinter>=5.2.2
pyyaml>=6.0.1
av>=12.0.0
cyndilib>=0.1.0
numpy>=1.26.0
